requirements.txt
torch==1.6.0
torchvision==0.7.0
tensorboard==2.3.0
Pillow==7.2.0
Flask==1.1.2
requests==2.24.0
